backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
SQLAlchemy==2.0.35
psycopg2==2.9.9
alembic==1.13.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
geoalchemy2==0.15.2
shapely==2.0.4
pytest==8.3.2
httpx==0.27.2
pytest-asyncio==0.23.8
